+++ R/extendr-wrappers.R
@@ -79,4 +79,10 @@ ferx_rust_inits_from_nca <- function(model_path, data_path, method) {
   .Call("wrap__ferx_rust_inits_from_nca", model_path, data_path, method)
 }
 
+#' @title Internal Rust backend binding
+#' @keywords internal
+ferx_rust_prepare_frem <- function(model_path, data_path, covariates, categorical_covariates, output_model_path, output_data_path) {
+  .Call("wrap__ferx_rust_prepare_frem", model_path, data_path, as.character(covariates), as.character(categorical_covariates), output_model_path, output_data_path)
+}
+
 # nolint end